Rubocop + linter correction
